+++ snowddl/parser/_yaml.py
@@ -1,6 +1,11 @@
 from pathlib import Path
 from yaml import SafeLoader, Node, add_constructor
 
+from snowddl.fernet.wrapper import FernetWrapper
+
+
+fernet_wrapper = FernetWrapper()
+
 
 class SnowDDLLoader(SafeLoader):
     pass
@@ -25,4 +30,16 @@ def include_constructor(loader: SnowDDLLoader, node: Node):
     return include_path.read_text(encoding="utf-8")
 
 
+def encrypt_constructor(loader: SnowDDLLoader, node: Node):
+    yaml_path = Path(loader.name)
+
+    raise ValueError(f"Detected non-encrypted value in file [{yaml_path}]")
+
+
+def decrypt_constructor(loader: SnowDDLLoader, node: Node):
+    return fernet_wrapper.decrypt(str(node.value))
+
+
 add_constructor("!include", include_constructor, Loader=SnowDDLLoader)
+add_constructor("!encrypt", encrypt_constructor, Loader=SnowDDLLoader)
+add_constructor("!decrypt", decrypt_constructor, Loader=SnowDDLLoader)